Made the installation of rpms conditional to the CI environment.
The installation in the buildtree environment is not useful as all is pre-installed in the image.

+++ continuousintegration/docker/ctafrontend/cc7/opt/run/bin/ctaeos-mgm.sh
@@ -2,13 +2,15 @@
 
 . /opt/run/bin/init_pod.sh
 
-yum-config-manager --enable cta-artifacts
-yum-config-manager --enable eos-citrine-commit
-yum-config-manager --enable eos-citrine-depend
-yum-config-manager --enable eos-citrine
+if [ ! -e /etc/buildtreeRunner ]; then
+  yum-config-manager --enable cta-artifacts
+  yum-config-manager --enable eos-citrine-commit
+  yum-config-manager --enable eos-citrine-depend
+  yum-config-manager --enable eos-citrine
 
-# Install missing RPMs
-yum -y install eos-client eos-server xrootd-client xrootd-debuginfo xrootd-server cta-cli cta-debuginfo sudo
+  # Install missing RPMs
+  yum -y install eos-client eos-server xrootd-client xrootd-debuginfo xrootd-server cta-cli cta-debuginfo sudo
+fi
 
 # create local users as the mgm is the only one doing the uid/user/group mapping in the full infrastructure
 groupadd --gid 1100 eosusers
