+++ continuousintegration/docker/ctafrontend/cc7/opt/run/bin/ctafrontend.sh
@@ -2,12 +2,14 @@
 
 . /opt/run/bin/init_pod.sh
 
+if [ ! -e /etc/buildtreeRunner ]; then
 yum-config-manager --enable cta-artifacts
 yum-config-manager --enable ceph
 
 # Install missing RPMs
 # cta-catalogueutils is needed to delete the db at the end of instance
 yum -y install cta-frontend cta-debuginfo cta-catalogueutils ceph-common
+fi
 
 yes | cp -r /opt/ci/ctafrontend/etc /
